CleanStrike/src/main/java/com/amaap/cleanstrike/controller/CarromBoardController.java
package com.amaap.cleanstrike.controller;

import com.amaap.cleanstrike.domain.model.CarromBoard;
import com.amaap.cleanstrike.domain.model.exception.InvalidCarromBoardDataException;
import com.amaap.cleanstrike.service.CarromBoardService;

public class CarromBoardController {
    CarromBoardService carromBoardService;

    public CarromBoardController(CarromBoardService carromBoardService) {
        this.carromBoardService = carromBoardService;
    }

    public CarromBoard create(int numberOfBlackCoins, int numberOfRedCoins) throws InvalidCarromBoardDataException {
        return carromBoardService.create(numberOfBlackCoins,numberOfRedCoins);
    }
}
